added WIP tag_scraper & added raw_tag datapoint to mongo documents
currently still WIP and working only with fake data, since the azure package we have is limited to 5k requests per month, so its better to test before using it
there needs to be an exeption for articles that have a article lenght of over 5k chars, since thats the limit for a single processable document by azure. meaning we'll have  to spilt those in two or more parts. unfortunately that also means that there will be even more requests to the azure service which could become problematic with our package.